Treat review comments
